dashboard/requirements.txt
dash-bootstrap-components==0.10.4
dash-core-components==1.10.2
dash-daq==0.5.0
dash-devices==0.1.3
dash-html-components==1.0.3
dash-renderer==1.6.0
dash-table==4.9.0
paho-mqtt==1.5.0
pickleDB==0.9.2
quart==0.13.0
quart-compress==0.2.1
